Add BotService class
