--- tg_bot/messaging/__init__.py
@@ -1,7 +1,0 @@
-__all__ = (
-    "send_files",
-    "send_message",
-)
-
-from .send_files import send_files
-from .send_message import send_message
